+++ ts/src/sdk/provider/createProviderSDK.server.ts
@@ -1,6 +1,6 @@
 import { createSDK } from "../../generated/createProviderSDK.ts";
-import { createGrpcTransport } from "../../transport/grpc/createGrpcTransport.ts";
 import type { PickByPath } from "../../utils/types.ts";
+import { createGrpcTransport } from "../transport/grpc/createGrpcTransport.ts";
 
 export type { PayloadOf, ResponseOf } from "../types.ts";
