+++ ts/src/sdk/transport/transportUtils.ts
